# internal/httpapi/openapi.yaml
openapi: 3.0.3
info:
  title: CostScope Cost API
  description: Read-only API over ClickHouse view `oci_cost_report_attributed`.
  version: 1.0.0
servers:
  - url: /
...
paths:
...
  /v1/costs/summary:
    get:
      tags: [costs]
      summary: Total cost summary for the filtered window
      parameters:
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/env"
        - $ref: "#/components/parameters/cost_center"
        - $ref: "#/components/parameters/component_type"
        - $ref: "#/components/parameters/compartment"
        - $ref: "#/components/parameters/service"
        - $ref: "#/components/parameters/resource_type"
        - $ref: "#/components/parameters/resource_name"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "400":
          $ref: "#/components/responses/Validation"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/costs/exec-summary:
    get:
      tags: [costs]
      summary: Aggregate payload for the executive summary page (one round trip)
      description: >
        Returns summary totals, monthly timeseries, cost-center and environment
        breakdowns, top-N breakdown for `dimension`, a monthly series per top-N
        dimension value, and freshness (best-effort, null on failure) in one
        `data` object. Sub-queries run concurrently server-side.
      parameters:
        - name: dimension
          in: query
          schema:
            type: string
            enum: [service, compartment, environment, cost_center, component_type, resource_type, resource_name]
            default: cost_center
        - name: top
          in: query
          description: Number of per-value monthly series to return
          schema: {type: integer, default: 7, minimum: 1, maximum: 20}
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/env"
        - $ref: "#/components/parameters/cost_center"
        - $ref: "#/components/parameters/component_type"
        - $ref: "#/components/parameters/compartment"
        - $ref: "#/components/parameters/service"
        - $ref: "#/components/parameters/resource_type"
        - $ref: "#/components/parameters/resource_name"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "400":
          $ref: "#/components/responses/Validation"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/costs/timeseries:
    get:
      tags: [costs]
      summary: Cost over time
      parameters:
        - name: granularity
          in: query
          schema:
            type: string
            enum: [hour, day, week, month]
            default: day
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/env"
        - $ref: "#/components/parameters/cost_center"
        - $ref: "#/components/parameters/component_type"
        - $ref: "#/components/parameters/compartment"
        - $ref: "#/components/parameters/service"
        - $ref: "#/components/parameters/resource_type"
        - $ref: "#/components/parameters/resource_name"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "400":
          $ref: "#/components/responses/Validation"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/costs/anomalies:
    get:
      tags: [costs]
      summary: Daily cost anomalies per dimension value (robust MAD z-score)
      description: >
        Flags days whose cost deviates from the trailing `window`-day median by a
        robust z-score (median absolute deviation based). Rows include baseline,
        deviation, z_score, severity (warning ≥ min_z, critical ≥ 5) and direction
        (spike/drop). The current, partial day is excluded.
      parameters:
        - name: dimension
          in: query
          schema:
            type: string
            enum: [service, compartment, environment, cost_center, component_type, resource_type, resource_name]
            default: service
        - name: window
          in: query
          description: Trailing baseline window in days
          schema: {type: integer, default: 28, minimum: 7, maximum: 90}
        - name: min_z
          in: query
          description: Minimum absolute robust z-score to report
          schema: {type: number, default: 3, minimum: 1, maximum: 20}
        - name: min_impact
          in: query
          description: Minimum absolute deviation from baseline (currency units)
          schema: {type: number, default: 50, minimum: 0}
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/env"
        - $ref: "#/components/parameters/cost_center"
        - $ref: "#/components/parameters/component_type"
        - $ref: "#/components/parameters/compartment"
        - $ref: "#/components/parameters/service"
        - $ref: "#/components/parameters/resource_type"
        - $ref: "#/components/parameters/resource_name"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "400":
          $ref: "#/components/responses/Validation"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/costs/trends:
    get:
      tags: [costs]
      summary: Period-over-period trend per dimension value
      description: >
        Compares the requested period against the equal-length period immediately
        before it. Rows include current_cost, previous_cost, change_amount,
        change_pct, slope (fitted cost change per day over the current period) and
        direction (rising/falling/flat/new/gone; flat when the change is under 5%).
      parameters:
        - name: dimension
          in: query
          schema:
            type: string
            enum: [service, compartment, environment, cost_center, component_type, resource_type, resource_name]
            default: service
        - name: granularity
          in: query
          description: >
            Time-axis resolution of the slope fit only; the current/previous
            period split is always day-exact, so unaligned starts are safe.
            slope remains a per-day rate.
          schema:
            type: string
            enum: [day, week, month]
            default: day
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/env"
        - $ref: "#/components/parameters/cost_center"
        - $ref: "#/components/parameters/component_type"
        - $ref: "#/components/parameters/compartment"
        - $ref: "#/components/parameters/service"
        - $ref: "#/components/parameters/resource_type"
        - $ref: "#/components/parameters/resource_name"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "400":
          $ref: "#/components/responses/Validation"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/costs/breakdown:
    get:
      tags: [costs]
      summary: Cost grouped by one dimension
      parameters:
        - name: dimension
          in: query
          schema:
            type: string
            enum: [service, compartment, environment, cost_center, component_type, resource_type, resource_name]
            default: service
        - name: limit
          in: query
          schema: {type: integer, default: 20, minimum: 1, maximum: 100}
        - name: series
          in: query
          description: Include a cost series for each returned row. Defaults to false.
          schema: {type: boolean, default: false}
        - $ref: "#/components/parameters/granularity"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/env"
        - $ref: "#/components/parameters/cost_center"
        - $ref: "#/components/parameters/component_type"
        - $ref: "#/components/parameters/compartment"
        - $ref: "#/components/parameters/service"
        - $ref: "#/components/parameters/resource_type"
        - $ref: "#/components/parameters/resource_name"
      responses:
        "200":
          description: Breakdown envelope. `series` is omitted unless `series=true`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BreakdownEnvelope"
        "400":
          $ref: "#/components/responses/Validation"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/costs/resources:
    get:
      tags: [costs]
      summary: Paginated per-resource costs
      parameters:
        - name: page
          in: query
          schema: {type: integer, default: 1, minimum: 1, maximum: 100000}
        - name: limit
          in: query
          schema: {type: integer, default: 50, minimum: 1, maximum: 500}
        - name: sort
          in: query
          schema: {type: string, default: cost}
        - name: direction
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/env"
        - $ref: "#/components/parameters/cost_center"
        - $ref: "#/components/parameters/component_type"
        - $ref: "#/components/parameters/compartment"
        - $ref: "#/components/parameters/service"
        - $ref: "#/components/parameters/resource_type"
        - $ref: "#/components/parameters/resource_name"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "400":
          $ref: "#/components/responses/Validation"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/costs/resources/grouped:
    get:
      tags: [costs]
      summary: Lazily expandable grouped resource-month costs
      description: >
        Groups a USD-only resource-month dataset. `group1_value` and `group2_value`
        scope lazy expansion; `q` is a case-insensitive resource-text search applied
        before grouping. A synthetic `other` row represents children past the top-15 cap.
      parameters:
        - name: group1
          in: query
          required: true
          schema:
            type: string
            enum: [service, compartment, environment, cost_center, component_type, resource_type, resource_name, period]
        - name: group2
          in: query
          schema:
            type: string
            enum: [service, compartment, environment, cost_center, component_type, resource_type, resource_name, period]
        - name: group1_value
          in: query
          description: Parent group value; use `__untagged__` for blank tag dimensions.
          schema: {type: string}
        - name: group2_value
          in: query
          description: Second parent group value; requires `group1_value` and `group2`.
          schema: {type: string}
        - name: q
          in: query
          description: Case-insensitive substring search across resource text columns.
          schema: {type: string}
        - name: hide_zero
          in: query
          description: When true, drops groups/leaves whose rounded cost is $0 ("hide noise").
          schema: {type: boolean, default: false}
        - name: grain
          in: query
          description: Required leaf grain; only month is supported.
          schema: {type: string, enum: [month], default: month}
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/env"
        - $ref: "#/components/parameters/cost_center"
        - $ref: "#/components/parameters/component_type"
        - $ref: "#/components/parameters/compartment"
        - $ref: "#/components/parameters/service"
        - $ref: "#/components/parameters/resource_type"
        - $ref: "#/components/parameters/resource_name"
      responses:
        "200":
          description: Group, synthetic remainder, or resource-month leaf rows in an envelope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupedResourcesEnvelope"
        "400":
          $ref: "#/components/responses/Validation"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/costs/resources/{ocid}:
    get:
      tags: [costs]
      summary: Cost detail for a single resource OCID
      description: Honors the shared date and dimension filters in addition to the required path OCID.
      parameters:
        - name: ocid
          in: path
          required: true
          schema: {type: string, maxLength: 512}
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/env"
        - $ref: "#/components/parameters/cost_center"
        - $ref: "#/components/parameters/component_type"
        - $ref: "#/components/parameters/compartment"
        - $ref: "#/components/parameters/service"
        - $ref: "#/components/parameters/resource_type"
        - $ref: "#/components/parameters/resource_name"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "400":
          $ref: "#/components/responses/Validation"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/costs/lineitems:
    get:
      tags: [costs]
      summary: Bucketed cost detail under the supplied shared filters (resource_name or ocid optionally narrow to one resource)
      parameters:
        - name: resource_name
          in: query
          schema: {type: string}
        - name: ocid
          in: query
          schema: {type: string, maxLength: 512}
        - $ref: "#/components/parameters/lineitem_granularity"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/env"
        - $ref: "#/components/parameters/cost_center"
        - $ref: "#/components/parameters/component_type"
        - $ref: "#/components/parameters/compartment"
        - $ref: "#/components/parameters/service"
        - $ref: "#/components/parameters/resource_type"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "400":
          $ref: "#/components/responses/Validation"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/costs/filters:
    get:
      tags: [costs]
      summary: Distinct filter values for the current window
      parameters:
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/env"
        - $ref: "#/components/parameters/cost_center"
        - $ref: "#/components/parameters/component_type"
        - $ref: "#/components/parameters/compartment"
        - $ref: "#/components/parameters/service"
        - $ref: "#/components/parameters/resource_type"
        - $ref: "#/components/parameters/resource_name"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "400":
          $ref: "#/components/responses/Validation"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/costs/freshness:
    get:
      tags: [ops]
      summary: Latest data timestamp in the source view
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "503":
          $ref: "#/components/responses/Unavailable"
components:
  parameters:
    granularity:
      name: granularity
      in: query
      description: Bucket size for an optional breakdown series
      schema:
        type: string
        enum: [hour, day, week, month]
        default: day
    lineitem_granularity:
      name: granularity
      in: query
      description: Bucket size for the rollup rows (cost, my_cost, line_items, overage_items per bucket)
      schema:
        type: string
        enum: [day, week, month]
        default: day
    start:
      name: start
      in: query
      description: RFC3339 window start (default now-1 month)
      schema: {type: string, format: date-time}
    end:
      name: end
      in: query
      description: RFC3339 window end (default now); range max 400 days
      schema: {type: string, format: date-time}
    env:
      name: env
      in: query
      description: Use `__untagged__` to select blank values.
      schema: {type: string}
    cost_center:
      name: cost_center
      in: query
      description: Use `__untagged__` to select blank values.
      schema: {type: string}
    component_type:
      name: component_type
      in: query
      description: Use `__untagged__` to select blank values.
      schema: {type: string}
    compartment:
      name: compartment
      in: query
      description: Use `__untagged__` to select blank values.
      schema: {type: string}
    service:
      name: service
      in: query
      description: Use `__untagged__` to select blank values.
      schema: {type: string}
    resource_type:
      name: resource_type
      in: query
      description: Use `__untagged__` to select blank values.
      schema: {type: string}
    resource_name:
      name: resource_name
      in: query
      description: Canonical resource-name value. Blank name tags resolve to `untagged · <product_service> · …<OCID tail>`; `__untagged__` does not apply to this parameter.
      schema: {type: string}
  schemas:
    Envelope:
      type: object
      properties:
        data:
          nullable: true
        meta:
          type: object
          additionalProperties: true
        error:
          nullable: true
          type: object
          properties:
            code: {type: string}
            message: {type: string}
    BreakdownEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: array
              items:
                type: object
                required: [dimension_value, currency, cost, resources]
                properties:
                  dimension_value: {type: string}
                  currency: {type: string}
                  cost:
                    oneOf:
                      - type: number
                        format: double
                      - type: string
                    description: >-
                      Number in the default mode; decimal string (e.g. "39957.95")
                      when `series=true` — series-mode costs are decimal strings to
                      avoid float precision loss, parse only at display.
                  resources: {type: integer}
                  series:
                    type: array
                    description: Present only when `series=true`.
                    items:
                      type: object
                      required: [date, cost]
                      properties:
                        date: {type: string, format: date-time}
                        cost:
                          type: string
                          description: Decimal string, never a float.
    GroupedResourcesEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: array
              items:
                oneOf:
                  - $ref: "#/components/schemas/GroupedResourceGroup"
                  - $ref: "#/components/schemas/GroupedResourceLeaf"
                discriminator:
                  propertyName: kind
    GroupedResourceGroup:
      type: object
      required: [kind, depth, group_value, currency, subtotal_cost, row_count]
      properties:
        kind: {type: string, enum: [group, other]}
        depth: {type: integer, minimum: 0}
        group_value: {type: string}
        currency: {type: string, enum: [USD]}
        subtotal_cost: {type: string, description: Decimal string.}
        row_count: {type: integer}
    GroupedResourceLeaf:
      type: object
      required: [kind, period, environment, cost_center, component_type, compartment, service, resource_type, resource_name, ocid, currency, cost]
      properties:
        kind: {type: string, enum: [leaf]}
        period: {type: string, example: "2026-07"}
        environment: {type: string}
        cost_center: {type: string}
        component_type: {type: string}
        compartment: {type: string}
        service: {type: string}
        resource_type: {type: string}
        resource_name: {type: string}
        ocid: {type: string}
        currency: {type: string, enum: [USD]}
        cost: {type: string, description: Decimal string.}
  responses:
    Ok:
      description: Success envelope; rows in `data`, freshness/pagination in `meta`
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    Validation:
      description: Invalid parameters (`error.code = VALIDATION_ERROR`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    Unavailable:
      description: ClickHouse unavailable (`error.code = UPSTREAM_ERROR` or `UNHEALTHY`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"